tsuki s

+++ includes/get_next_line.h
@@ -3,16 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aijadid <aijadid@student.1337.ma>          +#+  +:+       +#+        */
+/*   By: ayboudya <ayboudya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/15 22:12:30 by ayboudya          #+#    #+#             */
-/*   Updated: 2025/12/04 18:36:12 by aijadid          ###   ########.fr       */
+/*   Updated: 2025/12/18 14:35:58 by ayboudya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef GET_NEXT_LINE_H
 # define GET_NEXT_LINE_H
 # include <fcntl.h>
+# include <stdio.h>
 # include <stdlib.h>
 # include <unistd.h>
 # ifndef BUFFER_SIZE
@@ -24,7 +25,7 @@ char	*ft_strjoin(char *s1, char *s2);
 int		ft_strcmp(char *str1, char *str2);
 char	*ft_substr(char *s, unsigned int start, size_t len);
 char	*ft_strdup(char *str);
-int	ft_strlen(char *str);
+size_t	ft_strlen(char *str);
 char	*get_next_line(int fd);
 char	*test2(char *line);
 char	*test(char *line);
